Back to only effect-coding 2-level factors
I realized that there is a greater problem with factor variables being treated as numeric by sim_fixef. As a result, my previous code that effect-coded categorical variables with more than 2 levels is effectively ignored. 

This also means that the value.labels option for sim_factor doesn't work; a factor variable is not created and the value labels are converted back to numeric. Currently, this has to be the way it is, because the matrix multiplication of the beta weights with the predictor matrix (i.e., when creating the design matrix) fails if factor variables with more than two levels are treated as factors. This is a problem, though, because it means that factor levels are treated as integers (i.e., ordinal and equally-spaced). 

Also, I apologize for repeated commits. I have now developed a series of tests I can run locally, so the interim activity on this branch shouldn't be so high.

I intend to find a solution to this problem, but for now wanted to ef